test if s is null
sometimes the variable "s" is set to null when set_nl is called. This
is very bad and causes the program to crash often when dealing with
music with secondary layers such as piano music. Adding the immediate
return of s if it is null seems to fix the problem.